hadoop: Test failing load MR #277

+++ opencga-storage/opencga-storage-core/src/test/java/org/opencb/opencga/storage/core/variant/VariantStorageManagerTestUtils.java
@@ -76,8 +76,10 @@ public abstract class VariantStorageManagerTestUtils extends GenericTest impleme
     public static URI getResourceUri(String resourceName) throws IOException {
         Path rootDir = getTmpRootDir();
         Path resourcePath = rootDir.resolve(resourceName);
-        Files.copy(VariantStorageManagerTest.class.getClassLoader().getResourceAsStream(resourceName), resourcePath, StandardCopyOption
-                .REPLACE_EXISTING);
+        if (!resourcePath.toFile().exists()) {
+            Files.copy(VariantStorageManagerTest.class.getClassLoader().getResourceAsStream(resourceName), resourcePath, StandardCopyOption
+                    .REPLACE_EXISTING);
+        }
         return resourcePath.toUri();
     }
